% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AEO2019_dat}
\alias{AEO2019_dat}
\title{AEO2019_dat}
\format{
A dataframe with 35 rows and 11 columns.
}
\usage{
AEO2019_dat
}
\description{
Year 2019: Industrial Sector Macroeconomic Indicators (Value of Shipments).
\describe{
\item{\code{Type}}{Type of sector. Manufacturing or Non-manufacturing}
\item{\code{Sector}}{kind of sector}
\item{\code{Years of projections}}{Years}
}
}
\keyword{datasets}
