% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AEO2023_dat}
\alias{AEO2023_dat}
\title{AEO2023_dat}
\format{
A dataframe with 35 rows and 32 columns.
}
\usage{
AEO2023_dat
}
\description{
Year 2023: Industrial Sector Macroeconomic Indicators (Value of Shipments).
\describe{
\item{\code{Type}}{Type of sector. Manufacturing or Non-manufacturing}
\item{\code{Sector}}{kind of sector}
\item{\code{Years of projections}}{Years}
}
}
\keyword{datasets}
