% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_view.R
\name{branch}
\alias{branch}
\title{Move through a series of branches (subcategories) in your EIA tree (category)}
\usage{
branch(tree, branch_names)
}
\arguments{
\item{tree}{An EIA tree (category)}

\item{branch_names}{A list of branch (subcategory) names}
}
\description{
Find possible branches with \code{\link{branches}}
}
