% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_view.R
\name{branch_ID}
\alias{branch_ID}
\title{Find the ID number of a branch (subcategory)}
\usage{
branch_ID(tree)
}
\arguments{
\item{tree}{An EIA tree}
}
\description{
Find the ID number of a branch (subcategory)
}
