% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_view.R
\name{branches}
\alias{branches}
\title{Find available branches (subcategories) in your EIA tree (category)}
\usage{
branches(tree)
}
\arguments{
\item{tree}{An EIA tree  (category)}
}
\description{
Find available branches (subcategories) in your EIA tree (category)
}
