% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cooling}
\alias{cooling}
\title{Cooling}
\format{
A data frame with twenty two variables:
\describe{
\item{\code{Year}}{}
\item{\code{Month}}{}
\item{\code{PlantID}}{}
\item{\code{CoolingID}}{}
\item{\code{CoolingType}}{}
\item{\code{CoolingStatus}}{}
\item{\code{HoursInService}}{}
\item{\code{Chlorine (thousand lbs)}}{}
\item{\code{Diversion (gpm)}}{}
\item{\code{Withdrawal (gpm)}}{}
\item{\code{Discharge (gpm)}}{}
\item{\code{Consumption (gpm)}}{}
\item{\code{FlowMethod}}{}
\item{\code{IntakeAvgTemp (F)}}{}
\item{\code{IntakeMaxTemp}}{}
\item{\code{DischargeAvgTemp}}{}
\item{\code{DischargeMaxTemp}}{}
\item{\code{TempMethod}}{}
\item{\code{Diversion (MGal)}}{}
\item{\code{Withdrawal (MGal)}}{}
\item{\code{Discharge (MGal)}}{}
\item{\code{Consumption (MGal)}}{}
}
}
\usage{
cooling
}
\description{
Monthly entries for cooling of electricity generation plants over the years 2012-2020.
From EIA Form 923, Schedule 8.
}
\keyword{datasets}
