% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{generators}
\alias{generators}
\title{Generators}
\format{
A data frame with seventy five variables:
\describe{
\item{\code{Owner of Plants Reported on Form?}}{}
\item{\code{Other Relationships with Plants Reported on Form?}}{}
\item{\code{Entity Type}}{}
\item{\code{Year}}{Year}
}
}
\usage{
generators
}
\description{
Electric power generators in EIA form 860 over the years 2013-2017
}
\keyword{datasets}
