% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{join_and_drop}
\alias{join_and_drop}
\title{left join two dataframes, dropping extra copied columns}
\usage{
join_and_drop(df1, df2, by = NULL, rename = FALSE)
}
\arguments{
\item{df1}{a dataframe}

\item{df2}{a dataframe}

\item{by}{a character vector of variables to join by}

\item{rename}{logical}
}
\description{
left join two dataframes, dropping extra copied columns
}
