% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_view.R
\name{leaf}
\alias{leaf}
\title{Open a leaf (dataset) in your EIA tree (category)}
\usage{
leaf(tree, leaf_name)
}
\arguments{
\item{tree}{An EIA tree  (category)}

\item{leaf_name}{The name of the leaf (dataset) to extract}
}
\description{
Find possible leaves (datasets) with \code{\link{leaves}}
}
