% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_view.R
\name{leaf_info}
\alias{leaf_info}
\title{Find extra information about a leaf (dataset) in your tree (category)}
\usage{
leaf_info(tree, leaf_name)
}
\arguments{
\item{tree}{An EIA tree (category)}

\item{leaf_name}{A leaf (dataset) in your tree (category)}
}
\description{
Find extra information about a leaf (dataset) in your tree (category)
}
