% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_view.R
\name{leaves}
\alias{leaves}
\title{Find available leaves (datasets) in your EIA tree (category)}
\usage{
leaves(tree)
}
\arguments{
\item{tree}{An EIA tree (category)}
}
\description{
Find available leaves (datasets) in your EIA tree (category)
}
