% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{monthyear_to_date}
\alias{monthyear_to_date}
\title{Convert (and drop) month and year columns to a date column}
\usage{
monthyear_to_date(
  df,
  monthcol = "Month",
  yearcol = "Year",
  datecol = "Date",
  dateFormat = "\%m \%Y"
)
}
\arguments{
\item{df}{a dataframe with a column for month and a column for year.}

\item{monthcol}{length 1 character vector name of the month column.}

\item{yearcol}{length 1 character vector name of the year column.}

\item{datecol}{length 1 character vector name of the year column.}
}
\description{
Convert month and year to date
}
