% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{net_generation}
\alias{net_generation}
\title{Net Generation}
\format{
A data frame with twenty seven variables:
\describe{
\item{\code{PlantID}}{EIA-assigned plant code}
\item{\code{Combined Heat and Power}}{Indicates whether the generator is associated with a combined heat and power system (Yes/No)}
\item{\code{Plant}}{Plant name}
\item{\code{Operator}}{Is the reporting entity an operator of power plants reported on Schedule 2 of the form?}
\item{\code{OperatorID}}{ID of operator}
\item{\code{State}}{State for owner}
\item{\code{CensusRegion}}{Census Region}
\item{\code{NERCRegion}}{North American Electric Reliability Corporation (NERC)region in which the plant operates}
\item{\code{NAICS}}{North American Industry Classification System (NAICS) code 'that best describes the primary purpose of the plant.Electric utility plants and
independent power producers whose primary purpose is generating electricity for sale will generally use code 22. For generators whose primary business is an industrial or
commercial process (e.g., paper mills, refineries, chemical plants, etc.) and for which generating electricity is a secondary purpose, use a code other than 22.}
\item{\code{SectorNumber}}{Plant-level sector number, designated by the primary purpose, regulatory status and plant-level combined heat and power status"
1 = Electric Utility
2 = Independent Power Producer, Non-Combined Heat and Power
3 = Independent Power Producer, Combined Heat and Power
4 = Commercial, Non-Combined Heat and Power
5 = Commercial, Combined Heat and Power
6 = Industrial, Non-Combined Heat and Power
7 = Industrial, Combined Heat and Power"
}
\item{\code{Sector}}{Plant-level sector name, designated by the primary purpose, regulatory status and plant-level combined heat and power status}
\item{\code{GeneratorID}}{Generator identification number}
\item{\code{PrimeMover}}{EIA assigned code for the prime mover (i.e. the engine, turbine, water wheel, or similar machine that drives an electric generator)
BA	Energy Storage, Battery
CE	Energy Storage, Compressed Air
CP	Energy Storage, Concentrated Solar Power
FW	Energy Storage, Flywheel
PS	Energy Storage, Reversible Hydraulic Turbine (Pumped Storage)
ES	Energy Storage, Other (specify in SCHEDULE 7)
ST	Steam Turbine, including nuclear, geothermal and solar steam (does not include combined cycle)
GT	Combustion (Gas) Turbine (does not include the combustion turbine part of a combined cycle; see code CT, below)
IC	Internal Combustion Engine (diesel, piston, reciprocating)
CA	Combined Cycle Steam Part
CT	Combined Cycle Combustion Turbine Part
CS	Combined Cycle Single Shaft (combustion turbine and steam turbine share a single generator)
CC	Combined Cycle Total Unit (use only for plants/generators that are in planning stage, for which specific generator details cannot be provided)
HA	Hydrokinetic, Axial Flow Turbine
HB	Hydrokinetic, Wave Buoy
HK	Hydrokinetic, Other (specify in SCHEDULE 7)
HY	Hydroelectric Turbine (includes turbines associated with delivery of water by pipeline)
BT	Turbines Used in a Binary Cycle (including those used for geothermal applications)
PV	Photovoltaic
WT	Wind Turbine, Onshore
WS	Wind Turbine, Offshore
FC	Fuel Cell
OT	Other (specify in SCHEDULE 7)
}
\item{\code{Jan ... Dec}}{}
\item{\code{YearToDate}}{}
\item{\code{Year}}{}
\item{\code{Respondent Frequency}}{}
\item{\code{BalancingAuthorityCode}}{The plant's balancing authority
}
}
}
\usage{
net_generation
}
\description{
Net monthly electricity generation in MWh over the years 2012-2020.
From EIA Form 923, Page 4.
}
\keyword{datasets}
