% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{paste_columns}
\alias{paste_columns}
\title{Create a column by pasting multiple columns together}
\usage{
paste_columns(df, columns, new_col = "new_col", collapse = "; ")
}
\arguments{
\item{df}{dataframe}

\item{columns}{columns}

\item{new_col}{the new column name}

\item{collapse}{string to separate the values.}
}
\description{
Create a column by pasting multiple columns together
}
