% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wngsr.R
\name{wngsr}
\alias{wngsr}
\title{A function to return the latest EIA Weekly Natural Gas Storage Report.}
\usage{
wngsr()
}
\description{
The function wngsr will return a list object containing the time
the report was run, and the contents of the US EIA's Weekly Natural Gas
Storage Report.  The function has no arguments.
}
