% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cooling}
\alias{cooling}
\title{Cooling}
\format{A data frame with twenty two variables:
\describe{
\item{\code{Owner of Plants Reported on Form?}}{}
\item{\code{Other Relationships with Plants Reported on Form?}}{}
\item{\code{Entity Type}}{}
\item{\code{Year}}{Year}
}}
\usage{
cooling
}
\description{
Net monthly electricity generation in MWh over the years 2012-2017.
}
\keyword{datasets}
