% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cooling_detail}
\alias{cooling_detail}
\title{Detailed cooling information}
\format{
A data frame with twenty seven variables:
\describe{
\item{\code{Owner of Plants Reported on Form?}}{}
\item{\code{Other Relationships with Plants Reported on Form?}}{}
\item{\code{Entity Type}}{}
\item{\code{Year}}{Year}
}
}
\usage{
cooling_detail
}
\description{
Net monthly electricity generation in MWh over the years 2012-2017.
}
\keyword{datasets}
