% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{generators}
\alias{generators}
\title{generators}
\format{
A dataframe with 169422 rows and 61 columns.
}
\usage{
generators
}
\description{
Annual entries for each electric power generators in EIA form 860 over the years 2013-2020

\describe{
\item{\code{Utility ID}, \code{Utility Name}}{Unique identification codes and names of the electricity utilities.}
\item{\code{Plant Code}, \code{Plant Name}}{Numeric codes and names of the power plants.}
\item{\code{State}}{}
\item{\code{County}}{}
\item{\code{Generator ID}}{The unique generator identification commonly used by plant management.}
\item{\code{Prime Mover}}{The prime mover codes.
  \tabular{ll}{
    Code \tab Description \cr
    BA \tab Energy Storage, Battery \cr
    CE \tab Energy Storage, Compressed Air \cr
    CP \tab Energy Storage, Concentrated Solar Power \cr
    FW \tab Energy Storage, Flywheel \cr
    PS \tab Energy Storage, Reversible Hydraulic Turbine (Pumped Storage) \cr
    ES \tab Energy Storage, Other (specify in SCHEDULE 7) \cr
    ST \tab Steam Turbine, including nuclear, geothermal and solar steam (does not include combined cycle) \cr
    GT \tab Combustion (Gas) Turbine (does not include the combustion turbine part of a combined cycle; see code CT, below) \cr
    IC \tab Internal Combustion Engine (diesel, piston, reciprocating) \cr
    CA \tab Combined Cycle Steam Part \cr
    CT \tab Combined Cycle Combustion Turbine Part \cr
    CS \tab Combined Cycle Single Shaft (combustion turbine and steam turbine share a single generator) \cr
    CC \tab Combined Cycle Total Unit (use only for plants/generators that are in planning stage, for which specific generator details cannot be provided) \cr
    HA \tab Hydrokinetic, Axial Flow Turbine \cr
    HB \tab Hydrokinetic, Wave Buoy \cr
    HK \tab Hydrokinetic, Other (specify in SCHEDULE 7) \cr
    HY \tab Hydroelectric Turbine (includes turbines associated with delivery of water by pipeline) \cr
    BT \tab Turbines Used in a Binary Cycle (including those used for geothermal applications) \cr
    PV \tab Photovoltaic \cr
    WT \tab Wind Turbine, Onshore \cr
    WS \tab Wind Turbine, Offshore \cr
    FC \tab Fuel Cell \cr
    OT \tab Other (specify in SCHEDULE 7) \cr
  }
}
\item{\code{Unit Code}}{ If this generator
operates as a single unit with another generator (including as a combined cycle unit),
enter a unique 4-character code for the unit. All generators that operate as a unit must
have the same unit code. Leave blank if this generator does not operate as a single unit
with another generator}
\item{\code{Ownership}}{Ownership for each generator using the following codes:
  \tabular{ll}{
    Code \tab Description \cr
    S \tab Single ownership by respondent \cr
    J \tab Jointly owned with another entity \cr
    W \tab Wholly owned by an entity other than respondent \cr
  }
}
\item{\code{Duct Burners}}{"Yes" if 1) the generator has a combined cycle prime mover code of "Combined Cycle Steam Part (CA)", "Combined Cycle Single Shaft (CS)," or "Combined Cycle Total Unit (CC)," and 2) if the unit has duct-burners for supplementary firing of the turbine exhaust gas. Otherwise, check "No."}
\item{\code{Can Bypass Heat Recovery Steam Generator?}}{"Yes" if the generator has a combined cycle prime mover code of "Combined Cycle Combustion Turbine Part (CT)" or "Combined Cycle Total Unit (CC)" and 2) the combustion turbine can operate while bypassing the heat recovery steam generator. Otherwise, "No."}
\item{\code{RTO/ISO LMP Node Designation}}{
this generator operates in an electric system operated by a Regional Transmission
Organization (RTO) or Independent System Operator (ISO) and the RTO/ISO calculates
a nodal Locational Marginal Price (LMP) at the generator location, then provide the nodal
designation used to identify the price node in RTO/ISO LMP price reports.
}
\item{\code{RTO/ISO Location Designation for Reporting Wholesale Sales Data to FERC}}{
 If this generator operates in an electric
system operated by a Regional Transmission Organization (RTO) or Independent
System Operator (ISO) and the generator’s wholesale sales transaction data is reported
to FERC for the Electric Quarterly Report, then provide the designation used to report the
specific location of the wholesale sales transactions to FERC.}
\item{\code{Nameplate Capacity (MW)}}{
the highest value on the generator nameplate in MW rounded to the nearest tenth, as measured in
alternating current (AC). If the nameplate capacity is expressed in kilovolt amperes
(kVA), first convert the nameplate capacity to kilowatts by multiplying the corresponding
power factor by the kVA and then convert to megawatts by dividing by 1,000. Round this
value to the nearest tenth.}
\item{\code{Nameplate Power Factor}}{ The power factor stamped on the generator nameplate. This should be the same power factor used
to convert the generator’s kilovolt-ampere rating (kVA) to megawatts (MW) as directed for
line 1a above. Solar photovoltaic systems, wind turbines, batteries, fuel cells, and
flywheels may skip this question.}
\item{\code{Summer Capacity (MW)}}{The net summer capacity}
\item{\code{Winter Capacity (MW)}}{The net winter capacity}
\item{\code{Minimum Load (MW)}}{Minimum load (MW) at which the unit can operate continuously}
\item{\code{Uprate or Derate Completed During Year}}{“Yes” if an uprate or derate project was implemented during the
reporting year. “No” if it was not. If both an uprate and derate were implemented
during the reporting year}
\item{\code{Month Uprate or Derate Completed}}{}
\item{\code{Year Uprate or Derate Completed}}{}
\item{\code{Status}}{Operating and service statuses:
  \tabular{ll}{
    Code \tab Description \cr
    OP \tab Operating – in service (commercial operation) and producing some electricity. Includes peaking units that are run on an as needed (intermittent or seasonal) basis. \cr
    SB \tab Standby/Backup – available for service but not normally used (has little or no generation during the year) for this reporting period. \cr
    OS \tab Out of service – was not used for some or all of the reporting period and is NOT expected to be returned to service in the next calendar year. \cr
    OA \tab Out of service – was not used for some or all of the reporting period but is expected to be returned to service in the next calendar year. \cr
    RE \tab Retired – no longer in service and not expected to be returned to service. \cr
  }
}
\item{\code{Synchronized to Transmission Grid}}{If the status code is standby (SB),  “Yes” if the generator is currently
equipped to be synchronized to the grid when operating. “No” if it is not.}
\item{\code{Operating Month}}{Month the generator began commercial operation}
\item{\code{Operating Year}}{Year the generator began commercial operation}
\item{\code{Planned Retirement Month}}{Month the generator retired}
\item{\code{Planned Retirement Year}}{Year the generator retired}
\item{\code{Associated with Combined Heat and Power System}}{ “Yes” if this generator is associated with a combined heat and power system. “No” if it is not}
\item{\code{Sector Name}}{Plant-level sector number, designated by the primary purpose, regulatory status and plant-level combined heat and power status}
\item{\code{Sector}}{Plant-level sector name, designated by the primary purpose, regulatory status and plant-level combined heat and power status}
\item{\code{Topping or Bottoming}}{If the generator is associated with a combined heat and power system, indicates whether the generator is part of a topping cycle or a bottoming cycle'}
\item{\code{Energy Source 1, 2, 3, ... 6}}{The code representing the predominant (1-6) type of energy that fuels the generator}
\item{\code{Startup Source 1, 2, 3, ... 6}}{The code representing the first, second, third or fourth start-up and flame stabilization energy source used by the combustion unit(s) associated with this generator}
\item{\code{Solid Fuel Gasification System?}}{Indicates whether the generator is part of a solid fuel gasification system}
\item{\code{Carbon Capture Technology?}}{Indicates whether the generator uses carbon capture technology}
\item{\code{Turbines, Inverters, or Hydrokinetic Buoys}}{Number of wind turbines, or hydrokinetic buoys}
\item{\code{Time from Cold Shutdown to Full Load}}{The minimum amount of time required to bring the unit to full load from shutdown}
\item{\code{Fluidized Bed Technology?}}{Indicates whether the generator uses fluidized bed technology (Yes / No)}
\item{\code{Pulverized Coal Technology?}}{Indicates whether the generator uses pulverized coal technology(Yes / No)}
\item{\code{Stoker Technology?}}{Indicates whether the generator uses stoker technology (Yes / No)}
\item{\code{Other Combustion Technology?}}{Indicates whether the generator uses other combustion technologies (Yes / No)}
\item{\code{Subcritical Technology?}}{Indicates whether the generator uses subcritical technology (Yes / No)}
\item{\code{Supercritical Technology?}}{Indicates whether the generator uses supercritical technology (Yes / No)}
\item{\code{Ultrasupercritical Technology?}}{Indicates whether the generator uses ultra-supercritical technology (Yes/No)}
\item{\code{Planned Net Summer Capacity Uprate (MW)}}{Increase in summer capacity expected to be realized from the modification to the equipment}
\item{\code{Planned Net Winter Capacity Uprate (MW)}}{Increase in winter capacity expected to be realized from the uprate modification to the equipment}
\item{\code{Planned Uprate Month}}{Planned effective month that the generator is scheduled to enter operation after the uprate modification}
\item{\code{Planned Uprate Year}}{Planned effective year that the generator is scheduled to enter operation after the uprate modification}
\item{\code{Planned Net Summer Capacity Derate (MW)}}{Decrease in summer capacity expected to be realized from the derate modification to the equipment}
\item{\code{Planned Net Winter Capacity Derate (MW)}}{Decrease in winter capacity expected to be realized from the derate modification to the equipment}
\item{\code{Planned Derate Month}}{Planned effective month that the generator is scheduled to enter operation after the derate modification}
\item{\code{Planned Derate Year}}{Planned effective year that the generator is scheduled to enter operation after the derate modification}
\item{\code{Planned New Prime Mover}}{New prime mover for the planned repowered generator}
\item{\code{Planned Energy Source 1}}{New energy source code for the planned repowered generator}
\item{\code{Planned New Nameplate Capacity (MW)}}{The expected new namplate capacity for the generator}
\item{\code{Planned Repower Month}}{Planned effective month that the generator is scheduled to enter operation after the repowering is complete}
\item{\code{Planned Repower Year}}{Planned effective year that the generator is scheduled to enter operation after the repowering is complete}
\item{\code{Other Planned Modifications?}}{Indicates whether there are there other modifications planned for the generator}
\item{\code{Other Modifications Month}}{Planned effective month that the generator is scheduled to enter commercial operation after any other planned modification is complete}
\item{\code{Other Modifications Year}}{Planned effective year that the generator is scheduled to enter commercial operation after any other planned modofication is complete}
\item{\code{Cofire Fuels?}}{Can the generator co-fire fuels? (Yes / No)}
\item{\code{Switch Between Oil and Natural Gas?}}{Indicates whether the generator switch between oil and natural gas(Yes / No)}
\item{\code{Year}}{Year}
\item{\code{Technology}}{Generator Technology}
\item{\code{Turbines or Hydrokinetic Buoys}}{Number of wind turbines, or hydrokinetic buoys  }
\item{\code{Multiple Fuels?}}{Can the generator burn multiple fuels? (Yes/No)}
}
}
\keyword{datasets}
