% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{getCatEIA}
\alias{getCatEIA}
\title{A function to view the sub and parent categories of an Energy 
Information Administration (EIA) API data category.}
\usage{
getCatEIA(cat = 999999999, key)
}
\arguments{
\item{cat}{An EIA API data category number}

\item{key}{Your EIA API key, in quotes}
}
\description{
A function to view the sub and parent categories of a EIA API
data category. The function will return Series IDs in a category if present.
}
