% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{getEIA}
\alias{getEIA}
\title{A function to download data from the Energy Information Administration's (EIA's) API.}
\usage{
getEIA(ID, key)
}
\arguments{
\item{ID}{The EIA API Series ID for the data.}

\item{key}{Your EIA API key}
}
\description{
A function to download data from the EIA's API. Resulting time series are of class xts.
}
