% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{net_generation}
\alias{net_generation}
\title{Net Generation}
\format{
A data frame with twenty seven variables:
\describe{
\item{\code{PlantID}}{}
\item{\code{Combined Heat and Power}}{}
\item{\code{Plant}}{}
\item{\code{Operator}}{}
\item{\code{OperatorID}}{}
\item{\code{State}}{}
\item{\code{CensusRegion}}{}
\item{\code{NERCRegion}}{}
\item{\code{NAICS}}{}
\item{\code{SectorNumber}}{}
\item{\code{Sector}}{}
\item{\code{GeneratorID}}{}
\item{\code{PrimeMover}}{}
\item{\code{Jan ... Dec}}{}
\item{\code{YearToDate}}{}
\item{\code{Year}}{}
\item{\code{Respondent Frequency}}{}
\item{\code{BalancingAuthorityCode}}{}
}
}
\usage{
net_generation
}
\description{
Net monthly electricity generation in MWh over the years 2012-2020.
From EIA Form 923, Page 4.
}
\keyword{datasets}
