% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{utilities}
\alias{utilities}
\title{Utilities}
\format{A data frame with twelve variables:
\describe{
\item{\code{Owner of Plants Reported on Form?}}{}
\item{\code{Other Relationships with Plants Reported on Form?}}{}
\item{\code{Entity Type}}{}
\item{\code{Year}}{Year}
}

For further details, see \url{http://www.ssa.gov/oact/NOTES/as120/LifeTables_Body.html#wp1168594}}
\usage{
utilities
}
\description{
Electric utility information from EIA form 860 over the years 2013-2017.
}
\keyword{datasets}
