\name{wngsr}
\alias{wngsr}
\title{A function to return the latest EIA Weekly Natural
  Gas Storage Report.
}
\description{# The function wngsr will return a list object containing the time
  the report was run, and the contents of the US EIA's Weekly Natural Gas
  Storage Report.  The function has no arguments.
}
\usage{
wngsr()
}
\author{Matthew Brigida
}
\examples{
# To return the storage report simply run:

wngsr()

}
