\name{wngsr}
\alias{wngsr}
\title{A function to return the latest EIA Weekly Natural
  Gas Storage Report.
}
\description{# The function wngsr will return a list object containing (1) the time
the report was run, (2) the weeks the report covers, and (3) the
contents of the US EIA's Weekly Natural Gas Storage Report.  The
function has no arguments.
}
\usage{
wngsr()
}
\arguments{
}
\author{Matthew Brigida
}
\examples{
# To return the storage report simply run:

wngsr()

}
